+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name="smartgym.smartgym.SignIn"
+            android:name=".activities.SignIn"
             android:label="@string/app_name"
             android:configChanges="orientation"
             android:screenOrientation="portrait">
