+++ app/src/main/java/com/nwa/smartgym/models/HTTPResponse.java
@@ -1,4 +1,4 @@
-package smartgym.models;
+package com.nwa.smartgym.models;
 
 import java.util.Map;
