+++ app/src/main/res/layout/activity_sign_in.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="smartgym.smartgym.SignIn">
+    tools:context="activities.smartgym.SignIn">
 
     <!-- Login progress -->
     <ProgressBar
